fixup! regtest
